+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+- Fetch from /.well-known folder first #14
+
 ## [1.0.1(8)] - 2018-09-21
 ### Added
 - Support iOS 12 and Swift 4.2 #12